+++ src/lua/bindings/network.cpp
@@ -42,6 +42,15 @@ namespace lua::network
 
 	// Lua API: Function
 	// Table: network
+	// Name: is_session_started
+	// Returns true if the local player is in a multiplayer session.
+	static bool is_session_started()
+	{
+		return *big::g_pointers->m_gta.m_is_session_started;
+	}
+
+	// Lua API: Function
+	// Table: network
 	// Name: give_pickup_rewards
 	// Param: player: integer: Index of the player.
 	// Param: reward: integer: Index of the reward pickup.
@@ -171,11 +180,24 @@ namespace lua::network
 	// Lua API: Function
 	// Table: network
 	// Name: force_script_host
-	// Param: script_name: string: Name of the script
-	// Try to force ourself to be host for the given GTA Script.
+	// Param: script_name: string: Name of the script.
+	// Try to force ourself to be host for the given GTA Script. Needs to be called in the fiber pool or a loop.
 	static void force_script_host(const std::string& script_name)
 	{
 		big::scripts::force_host(rage::joaat(script_name));
+	}
+
+	// Lua API: function
+	// Table: script
+	// Name: force_script_on_player
+	// Param: player_idx: integer: Index of the player.
+	// Param: script_name: string: Name of the script.
+	// Param: script_name: integer: Instance ID of the script.
+	// Forces the given GTA script to be started on a player. Needs to be called in the fiber pool or a loop.
+	static void force_script_on_player(int player_idx, const std::string& script_name, int instance_id)
+	{
+		if (auto player = big::g_player_service->get_by_id(player_idx))
+			big::scripts::force_script_on_player(player, rage::joaat(script_name), instance_id);
 	}
 
 	// Lua API: Function
@@ -335,6 +357,7 @@ namespace lua::network
 		auto ns = state["network"].get_or_create<sol::table>();
 
 		ns["trigger_script_event"]                     = trigger_script_event;
+		ns["is_session_started"]                       = is_session_started;
 		ns["give_pickup_rewards"]                      = give_pickup_rewards;
 		ns["set_player_coords"]                        = set_player_coords;
 		ns["set_all_player_coords"]                    = set_all_player_coords;
@@ -345,6 +368,7 @@ namespace lua::network
 		ns["is_player_friend"]            = is_player_friend;
 		ns["get_flagged_modder_reason"]   = get_flagged_modder_reason;
 		ns["force_script_host"]           = force_script_host;
+		ns["force_script_on_player"]      = force_script_on_player;
 		ns["send_chat_message"]           = send_chat_message;
 		ns["send_chat_message_to_player"] = send_chat_message_to_player;
 		ns["get_player_rank"]             = get_player_rank;
